Improved Unpipelined Booth multiplier with higher operational frequency
The clock frequency in the un-pipelined booth multiplier has been increased from 100Mhz to roughly 150Mhz.

This is because in the earlier case, zero padding of the partial products was done one by one. This is un-necessary as we can do them all the same time. So this has resulted in improved clock frequency.